my_project_new_features_branch_s12

+++ cash-service/src/main/resources/db/migration/V2__create_transactions_table.sql
@@ -1,10 +1,10 @@
 CREATE TABLE IF NOT EXISTS cash_schema.cash_transactions (
                                                              id BIGSERIAL PRIMARY KEY,
                                                              user_login VARCHAR(50) NOT NULL,
-    amount INTEGER NOT NULL,
+    amount NUMERIC(19, 2) NOT NULL,
     action VARCHAR(10) NOT NULL CHECK (action IN ('PUT', 'GET')),
-    balance_before INTEGER NOT NULL,
-    balance_after INTEGER NOT NULL,
+    balance_before NUMERIC(19, 2) NOT NULL,
+    balance_after NUMERIC(19, 2) NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
     );
